# core/vm: Move interpreter.ReadOnly check into the opcode implementations