Fixes #23968.

I had trouble running tests locally, but `go test -v ./tests` passed (and displayed a lot of state tests lines). `make test` runs into various failures regarding "too many files open" or networking issues. In short: this change may be broken as I couldn't 100% verify with tests.