@holiman 2021-11-25
```
>>> build/cache/golangci-lint-1.42.0-linux-amd64/golangci-lint run --config .golangci.yml ./...
core/vm/interpreter.go:128: File is not `goimports`-ed (goimports)
			panic("readOnly execution requested on wrong chainRules");
```
There's not a lot of `;`-usage in golang :)