@holiman 2021-11-26
```
{"pc":0,"op":96,"gas":"0x989680","gasCost":"0x3","memory":"0x","memSize":0,"stack":[],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":2,"op":96,"gas":"0x98967d","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":4,"op":96,"gas":"0x98967a","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0","0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":6,"op":96,"gas":"0x989677","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":8,"op":97,"gas":"0x989674","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0","0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH2","error":""}
{"pc":11,"op":90,"gas":"0x989671","gasCost":"0x2","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0","0x0","0xff0b"],"returnData":"0x","depth":1,"refund":0,"opName":"GAS","error":""}
{"pc":12,"op":250,"gas":"0x98966f","gasCost":"0x963421","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0","0x0","0xff0b","0x98966f"],"returnData":"0x","depth":1,"refund":0,"opName":"STATICCALL","error":""}
{"pc":0,"op":96,"gas":"0x963165","gasCost":"0x3","memory":"0x","memSize":0,"stack":[],"returnData":"0x","depth":2,"refund":0,"opName":"PUSH1","error":""}
{"pc":2,"op":96,"gas":"0x963162","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x1"],"returnData":"0x","depth":2,"refund":0,"opName":"PUSH1","error":""}
{"pc":4,"op":85,"gas":"0x96315f","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x1","0x1"],"returnData":"0x","depth":2,"refund":0,"opName":"SSTORE","error":"write protection"}
{"pc":13,"op":0,"gas":"0x2624e","gasCost":"0x0","memory":"0x","memSize":0,"stack":["0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"STOP","error":""}
```
After
```
{"pc":0,"op":96,"gas":"0x989680","gasCost":"0x3","memory":"0x","memSize":0,"stack":[],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":2,"op":96,"gas":"0x98967d","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":4,"op":96,"gas":"0x98967a","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0","0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":6,"op":96,"gas":"0x989677","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH1","error":""}
{"pc":8,"op":97,"gas":"0x989674","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0","0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"PUSH2","error":""}
{"pc":11,"op":90,"gas":"0x989671","gasCost":"0x2","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0","0x0","0xff0b"],"returnData":"0x","depth":1,"refund":0,"opName":"GAS","error":""}
{"pc":12,"op":250,"gas":"0x98966f","gasCost":"0x963421","memory":"0x","memSize":0,"stack":["0x0","0x0","0x0","0x0","0xff0b","0x98966f"],"returnData":"0x","depth":1,"refund":0,"opName":"STATICCALL","error":""}
{"pc":0,"op":96,"gas":"0x963165","gasCost":"0x3","memory":"0x","memSize":0,"stack":[],"returnData":"0x","depth":2,"refund":0,"opName":"PUSH1","error":""}
{"pc":2,"op":96,"gas":"0x963162","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x1"],"returnData":"0x","depth":2,"refund":0,"opName":"PUSH1","error":""}
{"pc":4,"op":85,"gas":"0x96315f","gasCost":"0x4e20","memory":"0x","memSize":0,"stack":["0x1","0x1"],"returnData":"0x","depth":2,"refund":0,"opName":"SSTORE","error":""}
{"pc":13,"op":0,"gas":"0x2624e","gasCost":"0x0","memory":"0x","memSize":0,"stack":["0x0"],"returnData":"0x","depth":1,"refund":0,"opName":"STOP","error":""}

```
Diff: 
```diff
< {"pc":4,"op":85,"gas":"0x96315f","gasCost":"0x3","memory":"0x","memSize":0,"stack":["0x1","0x1"],"returnData":"0x","depth":2,"refund":0,"opName":"SSTORE","error":"write protection"}
---
> {"pc":4,"op":85,"gas":"0x96315f","gasCost":"0x4e20","memory":"0x","memSize":0,"stack":["0x1","0x1"],"returnData":"0x","depth":2,"refund":0,"opName":"SSTORE","error":""}
```
This may indeed be a bit problematic for the fuzzers to handle, since the error is never surfaced. I'll think about this a bit