@axic 2021-11-26
The gas *discrepancy* is still there, I wonder what gas value evmone outputs in most of these cases @chfast?