@holiman 2021-11-26
Yes, but the discrepancy is such that it was previously wrong, using the _previous_ value on error returns. And now it's more correct is all. I think we just didn't look to closely at the gas value whenever errors were involved.