@chfast 2021-11-26
> The gas _discrepancy_ is still there, I wonder what gas value evmone outputs in most of these cases @chfast?

I don't output `"gasCost"` at all because I don't know it. The tracing spec requires to report `"gasCost"` before execution of an instruction and in evmone gas cost computation and execution are mixed up.