@gumb0 2021-11-26
I'm getting these results from running evmone benchmarks with this against master.
```
name                                  old time/op  new time/op  delta
main/blake2b_shifts/2805nulls         15.9ms ± 1%  17.6ms ±15%    ~     (p=0.356 n=9+10)
main/blake2b_shifts/5610nulls         31.8ms ± 1%  30.7ms ± 1%  -3.22%  (p=0.000 n=10+10)
main/blake2b_shifts/8415nulls         51.3ms ± 7%  50.0ms ± 9%    ~     (p=0.247 n=10+10)
main/weierstrudel/0                    515µs ± 1%   496µs ± 0%  -3.71%  (p=0.000 n=10+10)
main/weierstrudel/1                   1.10ms ± 0%  1.07ms ± 0%  -3.12%  (p=0.000 n=10+10)
main/weierstrudel/3                   1.72ms ± 1%  1.66ms ± 1%  -3.81%  (p=0.000 n=10+10)
main/weierstrudel/9                   3.56ms ± 0%  3.39ms ± 1%  -4.73%  (p=0.000 n=9+9)
main/weierstrudel/14                  5.06ms ± 0%  4.81ms ± 0%  -5.01%  (p=0.000 n=9+9)
main/sha1_divs/empty                   239µs ± 0%   240µs ± 0%  +0.21%  (p=0.019 n=9+9)
main/sha1_divs/1351                   4.74ms ± 1%  4.71ms ± 1%  -0.76%  (p=0.000 n=10+9)
main/sha1_divs/2737                   9.21ms ± 1%  9.14ms ± 0%  -0.82%  (p=0.000 n=9+7)
main/sha1_divs/5311                   18.0ms ± 0%  17.9ms ± 1%  -0.63%  (p=0.001 n=7+10)
main/sha1_shifts/empty                 191µs ± 1%   181µs ± 0%  -4.82%  (p=0.000 n=9+10)
main/sha1_shifts/1351                 3.78ms ± 0%  3.60ms ± 1%  -4.91%  (p=0.000 n=10+10)
main/sha1_shifts/2737                 7.36ms ± 1%  7.00ms ± 1%  -4.82%  (p=0.000 n=9+9)
main/sha1_shifts/5311                 14.4ms ± 1%  13.7ms ± 0%  -4.73%  (p=0.000 n=9+8)
main/blake2b_huff/empty                116µs ± 1%   112µs ± 1%  -3.08%  (p=0.000 n=10+10)
main/blake2b_huff/2805nulls           2.16ms ± 0%  2.10ms ± 1%  -2.85%  (p=0.000 n=9+8)
main/blake2b_huff/5610nulls           4.17ms ± 0%  4.04ms ± 0%  -3.27%  (p=0.000 n=9+9)
main/blake2b_huff/8415nulls           6.09ms ± 0%  5.92ms ± 2%  -2.77%  (p=0.000 n=9+10)
micro/JUMPDEST_n0/empty               3.31ms ± 2%  3.17ms ± 0%  -4.20%  (p=0.000 n=10+10)
micro/jump_around/empty                163µs ± 0%   157µs ± 0%  -3.44%  (p=0.000 n=10+9)
micro/memory_grow_mstore/nogrow        483µs ± 1%   480µs ± 0%  -0.64%  (p=0.000 n=10+9)
micro/memory_grow_mstore/by1           493µs ± 1%   492µs ± 1%    ~     (p=0.247 n=10+10)
micro/memory_grow_mstore/by16          601µs ± 2%   607µs ± 1%  +0.95%  (p=0.027 n=10+8)
micro/memory_grow_mstore/by32          691µs ± 1%   693µs ± 1%    ~     (p=0.481 n=10+10)
micro/signextend/zero                  493µs ± 0%   488µs ± 1%  -0.87%  (p=0.000 n=9+9)
micro/signextend/one                   510µs ± 4%   495µs ± 0%  -2.84%  (p=0.000 n=10+10)
micro/loop_with_many_jumpdests/empty  3.83µs ± 3%  3.80µs ± 1%    ~     (p=0.530 n=9+9)
micro/memory_grow_mload/nogrow         249µs ± 1%   246µs ± 1%  -1.26%  (p=0.000 n=10+10)
micro/memory_grow_mload/by1            259µs ± 1%   255µs ± 1%  -1.41%  (p=0.000 n=10+10)
micro/memory_grow_mload/by16           343µs ± 1%   343µs ± 3%    ~     (p=0.579 n=10+10)
micro/memory_grow_mload/by32           448µs ± 0%   446µs ± 1%  -0.48%  (p=0.002 n=9+8)

name                                  old gas/s    new gas/s    delta
main/blake2b_shifts/2805nulls           215M ± 1%    197M ±14%    ~     (p=0.356 n=9+10)
main/blake2b_shifts/5610nulls           234M ± 1%    242M ± 1%  +3.33%  (p=0.000 n=10+10)
main/blake2b_shifts/8415nulls           236M ± 7%    242M ±10%    ~     (p=0.247 n=10+10)
main/weierstrudel/0                    92.3M ± 1%   95.9M ± 0%  +3.86%  (p=0.000 n=10+10)
main/weierstrudel/1                     118M ± 0%    122M ± 0%  +3.22%  (p=0.000 n=10+10)
main/weierstrudel/3                     117M ± 1%    122M ± 1%  +3.96%  (p=0.000 n=10+10)
main/weierstrudel/9                     117M ± 0%    123M ± 1%  +4.97%  (p=0.000 n=9+9)
main/weierstrudel/14                    118M ± 0%    124M ± 0%  +5.27%  (p=0.000 n=9+9)
main/sha1_divs/empty                    152M ± 0%    152M ± 0%  -0.21%  (p=0.019 n=9+9)
main/sha1_divs/1351                     166M ± 1%    167M ± 1%  +0.76%  (p=0.000 n=10+9)
main/sha1_divs/2737                     167M ± 1%    168M ± 0%  +0.83%  (p=0.000 n=9+7)
main/sha1_divs/5311                     166M ± 0%    168M ± 1%  +0.64%  (p=0.001 n=7+10)
main/sha1_shifts/empty                  185M ± 1%    194M ± 0%  +5.06%  (p=0.000 n=9+10)
main/sha1_shifts/1351                   201M ± 0%    211M ± 1%  +5.16%  (p=0.000 n=10+10)
main/sha1_shifts/2737                   202M ± 1%    212M ± 1%  +5.07%  (p=0.000 n=9+9)
main/sha1_shifts/5311                   202M ± 1%    212M ± 0%  +4.96%  (p=0.000 n=9+8)
main/blake2b_huff/empty                 121M ± 1%    124M ± 1%  +3.18%  (p=0.000 n=10+10)
main/blake2b_huff/2805nulls             142M ± 0%    146M ± 1%  +2.93%  (p=0.000 n=9+8)
main/blake2b_huff/5610nulls             144M ± 0%    149M ± 0%  +3.38%  (p=0.000 n=9+9)
main/blake2b_huff/8415nulls             145M ± 0%    149M ± 2%  +2.85%  (p=0.000 n=9+10)
micro/JUMPDEST_n0/empty                 160M ± 2%    167M ± 0%  +4.37%  (p=0.000 n=10+10)
micro/jump_around/empty                 302M ± 0%    312M ± 0%  +3.56%  (p=0.000 n=10+9)
micro/memory_grow_mstore/nogrow         127M ± 1%    128M ± 0%  +0.64%  (p=0.000 n=10+9)
micro/memory_grow_mstore/by1            125M ± 1%    126M ± 1%    ~     (p=0.247 n=10+10)
micro/memory_grow_mstore/by16           126M ± 2%    125M ± 1%  -0.94%  (p=0.027 n=10+8)
micro/memory_grow_mstore/by32           154M ± 1%    154M ± 1%    ~     (p=0.481 n=10+10)
micro/signextend/zero                   195M ± 0%    197M ± 1%  +0.88%  (p=0.000 n=9+9)
micro/signextend/one                    188M ± 4%    194M ± 0%  +2.88%  (p=0.000 n=10+10)
micro/loop_with_many_jumpdests/empty   8.36M ± 3%   8.41M ± 1%    ~     (p=0.530 n=9+9)
micro/memory_grow_mload/nogrow          230M ± 1%    233M ± 1%  +1.28%  (p=0.000 n=10+10)
micro/memory_grow_mload/by1             223M ± 1%    226M ± 1%  +1.42%  (p=0.000 n=10+10)
micro/memory_grow_mload/by16            209M ± 1%    209M ± 3%    ~     (p=0.579 n=10+10)
micro/memory_grow_mload/by32            229M ± 0%    230M ± 1%  +0.48%  (p=0.002 n=9+8)
```