@chfast 2021-11-26
My benchmarks on a desktop CPU.
```
name                                  old time/op  new time/op  delta                                                                                                                                                             
micro/memory_grow_mload/nogrow         249µs ± 1%   245µs ± 1%  -1.70%  (p=0.000 n=20+20)                                                                                                                                         
micro/memory_grow_mload/by1            257µs ± 1%   254µs ± 1%  -1.02%  (p=0.000 n=20+19)                                                                                                                                         
micro/memory_grow_mload/by16           321µs ± 1%   320µs ± 1%  -0.32%  (p=0.001 n=20+20)                                                                                                                                         
micro/memory_grow_mload/by32           398µs ± 1%   399µs ± 1%    ~     (p=0.142 n=20+20)                                                                                                                                         
micro/signextend/zero                  462µs ± 1%   457µs ± 1%  -0.97%  (p=0.000 n=20+20)                                                                                                                                         
micro/signextend/one                   474µs ± 1%   471µs ± 1%  -0.73%  (p=0.000 n=20+20)                                                                                                                                         
micro/loop_with_many_jumpdests/empty  3.40µs ± 2%  3.40µs ± 1%    ~     (p=0.934 n=20+17)                                                                                                                                         
micro/jump_around/empty                154µs ± 1%   152µs ± 1%  -1.35%  (p=0.000 n=20+20)                                                                                                                                         
micro/memory_grow_mstore/nogrow        485µs ± 0%   483µs ± 1%  -0.51%  (p=0.000 n=19+20)                                                                                                                                         
micro/memory_grow_mstore/by1           494µs ± 1%   492µs ± 1%  -0.36%  (p=0.000 n=19+20)                                                                                                                                         
micro/memory_grow_mstore/by16          574µs ± 1%   576µs ± 0%  +0.38%  (p=0.000 n=20+20)                                                                                                                                         
micro/memory_grow_mstore/by32          644µs ± 0%   646µs ± 0%  +0.33%  (p=0.000 n=20+20)                                                                                                                                         
micro/JUMPDEST_n0/empty               3.45ms ± 1%  3.49ms ± 0%  +1.35%  (p=0.000 n=19+20)                                                                                                                                         
main/sha1_divs/empty                   222µs ± 0%   221µs ± 0%  -0.74%  (p=0.000 n=20+19)                                                                                                                                         
main/sha1_divs/1351                   4.49ms ± 0%  4.47ms ± 1%  -0.56%  (p=0.000 n=19+20)
main/sha1_divs/2737                   8.73ms ± 0%  8.65ms ± 1%  -0.88%  (p=0.000 n=19+20)
main/sha1_divs/5311                   17.1ms ± 0%  16.9ms ± 0%  -0.95%  (p=0.000 n=18+19)
main/weierstrudel/0                    492µs ± 0%   486µs ± 0%  -1.28%  (p=0.000 n=19+20)
main/weierstrudel/1                   1.04ms ± 0%  1.03ms ± 0%  -0.97%  (p=0.000 n=19+20)
main/weierstrudel/3                   1.64ms ± 0%  1.61ms ± 1%  -1.69%  (p=0.000 n=19+20)
main/weierstrudel/9                   3.41ms ± 1%  3.32ms ± 1%  -2.83%  (p=0.000 n=19+20)
main/weierstrudel/14                  4.88ms ± 1%  4.73ms ± 1%  -3.06%  (p=0.000 n=20+19)
main/sha1_shifts/empty                 176µs ± 1%   179µs ± 1%  +1.68%  (p=0.000 n=20+19)
main/sha1_shifts/1351                 3.60ms ± 0%  3.67ms ± 0%  +1.89%  (p=0.000 n=18+20)
main/sha1_shifts/2737                 7.01ms ± 0%  7.15ms ± 1%  +2.02%  (p=0.000 n=19+20)
main/sha1_shifts/5311                 13.7ms ± 1%  14.0ms ± 1%  +2.38%  (p=0.000 n=20+19)
main/blake2b_huff/empty                109µs ± 0%   108µs ± 1%  -0.82%  (p=0.000 n=20+20)
main/blake2b_huff/2805nulls           2.08ms ± 0%  2.05ms ± 0%  -1.50%  (p=0.000 n=18+20)
main/blake2b_huff/5610nulls           4.04ms ± 0%  3.98ms ± 0%  -1.64%  (p=0.000 n=20+19)
main/blake2b_huff/8415nulls           5.90ms ± 0%  5.81ms ± 0%  -1.42%  (p=0.000 n=19+17)
main/blake2b_shifts/2805nulls         16.6ms ± 1%  16.4ms ± 1%  -1.00%  (p=0.000 n=18+19)
main/blake2b_shifts/5610nulls         33.1ms ± 1%  32.8ms ± 1%  -1.10%  (p=0.000 n=20+20)
main/blake2b_shifts/8415nulls         49.3ms ± 1%  48.7ms ± 1%  -1.13%  (p=0.000 n=19+19)
[Geo mean]                            1.30ms       1.29ms       -0.56%
```