@chfast 2021-11-29
New benchmarks after #23952  is merged.
```
name                                  old time/op  new time/op  delta                                                                                                                                                             
micro/memory_grow_mload/nogrow         237µs ± 1%   232µs ± 0%  -2.14%  (p=0.000 n=20+18)                                                                                                                                         
micro/memory_grow_mload/by1            245µs ± 1%   241µs ± 1%  -1.90%  (p=0.000 n=20+20)                                                                                                                                         
micro/memory_grow_mload/by16           310µs ± 0%   305µs ± 0%  -1.65%  (p=0.000 n=20+20)                                                                                                                                         
micro/memory_grow_mload/by32           388µs ± 0%   384µs ± 1%  -0.81%  (p=0.000 n=18+20)                                                                                                                                         
micro/signextend/zero                  444µs ± 1%   405µs ± 1%  -8.76%  (p=0.000 n=20+19)                                                                                                                                         
micro/signextend/one                   456µs ± 1%   417µs ± 0%  -8.45%  (p=0.000 n=20+19)                                                                                                                                         
micro/loop_with_many_jumpdests/empty  3.41µs ± 2%  3.40µs ± 1%    ~     (p=0.425 n=20+20)                                                                                                                                         
micro/jump_around/empty                149µs ± 1%   146µs ± 1%  -1.78%  (p=0.000 n=18+20)                                                                                                                                         
micro/memory_grow_mstore/nogrow        474µs ± 1%   467µs ± 1%  -1.53%  (p=0.000 n=20+20)                                                                                                                                         
micro/memory_grow_mstore/by1           483µs ± 1%   475µs ± 1%  -1.59%  (p=0.000 n=20+19)                                                                                                                                         
micro/memory_grow_mstore/by16          565µs ± 1%   557µs ± 0%  -1.41%  (p=0.000 n=20+19)                                                                                                                                         
micro/memory_grow_mstore/by32          636µs ± 1%   629µs ± 0%  -1.20%  (p=0.000 n=20+20)                                                                                                                                         
micro/JUMPDEST_n0/empty               3.13ms ± 0%  3.02ms ± 0%  -3.71%  (p=0.000 n=17+17)                                                                                                                                         
main/sha1_divs/empty                   207µs ± 0%   201µs ± 0%  -3.13%  (p=0.000 n=19+18)                                                                                                                                         
main/sha1_divs/1351                   4.16ms ± 0%  4.02ms ± 1%  -3.29%  (p=0.000 n=16+20)
main/sha1_divs/2737                   8.10ms ± 1%  7.82ms ± 1%  -3.44%  (p=0.000 n=19+18)
main/sha1_divs/5311                   15.8ms ± 1%  15.3ms ± 1%  -3.36%  (p=0.000 n=20+20)
main/weierstrudel/0                    473µs ± 0%   474µs ± 0%  +0.26%  (p=0.000 n=20+19)
main/weierstrudel/1                    987µs ± 0%   981µs ± 0%  -0.69%  (p=0.000 n=17+18)
main/weierstrudel/3                   1.55ms ± 1%  1.54ms ± 1%  -0.49%  (p=0.000 n=20+20)
main/weierstrudel/9                   3.19ms ± 1%  3.19ms ± 0%    ~     (p=0.443 n=18+20)
main/weierstrudel/14                  4.56ms ± 1%  4.55ms ± 1%    ~     (p=0.072 n=18+20)
main/sha1_shifts/empty                 162µs ± 1%   157µs ± 0%  -3.19%  (p=0.000 n=20+19)
main/sha1_shifts/1351                 3.26ms ± 0%  3.19ms ± 0%  -2.19%  (p=0.000 n=18+20)
main/sha1_shifts/2737                 6.36ms ± 0%  6.21ms ± 0%  -2.29%  (p=0.000 n=16+19)
main/sha1_shifts/5311                 12.4ms ± 1%  12.1ms ± 0%  -2.47%  (p=0.000 n=19+19)
main/blake2b_huff/empty                106µs ± 0%   104µs ± 0%  -1.91%  (p=0.000 n=19+18)
main/blake2b_huff/2805nulls           2.00ms ± 1%  1.96ms ± 1%  -2.08%  (p=0.000 n=20+20)
main/blake2b_huff/5610nulls           3.88ms ± 0%  3.79ms ± 1%  -2.30%  (p=0.000 n=18+18)
main/blake2b_huff/8415nulls           5.67ms ± 0%  5.55ms ± 0%  -2.13%  (p=0.000 n=19+18)
main/blake2b_shifts/2805nulls         15.7ms ± 0%  14.8ms ± 1%  -5.68%  (p=0.000 n=17+20)
main/blake2b_shifts/5610nulls         31.2ms ± 1%  29.5ms ± 1%  -5.63%  (p=0.000 n=20+20)
main/blake2b_shifts/8415nulls         46.5ms ± 1%  43.8ms ± 1%  -5.93%  (p=0.000 n=20+19)
[Geo mean]                            1.23ms       1.20ms       -2.60%
```